fix win cast

+++ source/s3_client.c
@@ -1910,7 +1910,7 @@ void s_acquire_mem_and_prepare_request(
         struct aws_s3_buffer_pool_reserve_meta meta = {
             .client = client,
             .meta_request = meta_request,
-            .size = aws_s3_request_get_part_size(request),
+            .size = (size_t)aws_s3_request_get_part_size(request),
         };
 
         struct aws_s3_reserve_memory_payload *payload =
